removing unused stuff
