+++ sdk/client_test.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
+	"errors"
 	"io"
 	"net/http"
 	"testing"
@@ -12,6 +13,7 @@ import (
 	"github.com/ONSdigital/dp-healthcheck/healthcheck"
 	dphttp "github.com/ONSdigital/dp-net/v2/http"
 	dpNetRequest "github.com/ONSdigital/dp-net/v2/request"
+	"github.com/ONSdigital/log.go/v2/log"
 	. "github.com/smartystreets/goconvey/convey"
 )
 
@@ -191,5 +193,44 @@ func TestHeaders(t *testing.T) {
 			So(request.Header, ShouldNotContainKey, dpNetRequest.CollectionIDHeaderKey)
 			So(request.Header, ShouldNotContainKey, dpNetRequest.DownloadServiceHeaderKey)
 		})
+	})
+}
+
+type mockReadCloser struct {
+	raiseError bool
+}
+
+// Implemented just to keep compiler happy for mock object
+func (m *mockReadCloser) Read(p []byte) (n int, err error) {
+	return 0, io.EOF
+}
+
+// Returns an error if `raiseError` is `true`, otherwise `false`
+func (m *mockReadCloser) Close() error {
+	if m.raiseError {
+		return errors.New("error closing body")
+	}
+	return nil
+}
+
+// Tests for `closeResponseBody` function
+func TestCloseResponseBody(t *testing.T) {
+	ctx := context.Background()
+	// Create a buffer to capture log output for tests
+	var buf bytes.Buffer
+	var fbBuf bytes.Buffer
+	log.SetDestination(&buf, &fbBuf)
+
+	Convey("Test function runs without logging an error if body.Close() completes without error", t, func() {
+		mockResponse := http.Response{Body: &mockReadCloser{raiseError: false}}
+
+		closeResponseBody(ctx, &mockResponse)
+		So(buf.String(), ShouldBeEmpty)
+	})
+	Convey("Test function logs an error if body.Close() returns an error", t, func() {
+		mockResponse := http.Response{Body: &mockReadCloser{raiseError: true}}
+
+		closeResponseBody(ctx, &mockResponse)
+		So(buf.String(), ShouldContainSubstring, "error closing http response body")
 	})
 }
